+++ ta4j-examples/src/main/java/ta4jexamples/indicators/CandlestickChart.java
@@ -70,11 +70,11 @@ public class CandlestickChart {
         for (int i = 0; i < nbBars; i++) {
             Bar bar = series.getBar(i);
             dates[i] = new Date(bar.getEndTime().toEpochSecond() * 1000);
-            opens[i] = bar.getOpenPrice().doubleValue();
-            highs[i] = bar.getHighPrice().doubleValue();
-            lows[i] = bar.getLowPrice().doubleValue();
-            closes[i] = bar.getClosePrice().doubleValue();
-            volumes[i] = bar.getVolume().doubleValue();
+            opens[i] = bar.getOpenPrice();
+            highs[i] = bar.getHighPrice();
+            lows[i] = bar.getLowPrice();
+            closes[i] = bar.getClosePrice();
+            volumes[i] = bar.getVolume();
         }
 
         return new DefaultHighLowDataset("btc", dates, highs, lows, opens, closes, volumes);
